// web/src/lib/api/models/SessionStatus.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Kimi Code CLI Web Interface
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Runtime status of a web session.
 * @export
 * @interface SessionStatus
 */
export interface SessionStatus {
    /**
     * Session unique ID
     * @type {string}
     * @memberof SessionStatus
     */
    sessionId: string;
    /**
     * Current session state
     * @type {string}
     * @memberof SessionStatus
     */
    state: SessionStatusStateEnum;
    /**
     * Monotonic sequence number
     * @type {number}
     * @memberof SessionStatus
     */
    seq: number;
    /**
     * 
     * @type {string}
     * @memberof SessionStatus
     */
    workerId?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SessionStatus
     */
    reason?: string | null;
    /**
     * 
     * @type {string}
     * @memberof SessionStatus
     */
    detail?: string | null;
    /**
     * Timestamp for this state
     * @type {Date}
     * @memberof SessionStatus
     */
    updatedAt: Date;
}


/**
 * @export
 */
export const SessionStatusStateEnum = {
    Stopped: 'stopped',
    Idle: 'idle',
    Busy: 'busy',
    Restarting: 'restarting',
    Error: 'error'
} as const;
export type SessionStatusStateEnum = typeof SessionStatusStateEnum[keyof typeof SessionStatusStateEnum];


/**
 * Check if a given object implements the SessionStatus interface.
 */
export function instanceOfSessionStatus(value: object): value is SessionStatus {
    if (!('sessionId' in value) || value['sessionId'] === undefined) return false;
    if (!('state' in value) || value['state'] === undefined) return false;
    if (!('seq' in value) || value['seq'] === undefined) return false;
    if (!('updatedAt' in value) || value['updatedAt'] === undefined) return false;
    return true;
}

export function SessionStatusFromJSON(json: any): SessionStatus {
    return SessionStatusFromJSONTyped(json, false);
}

export function SessionStatusFromJSONTyped(json: any, ignoreDiscriminator: boolean): SessionStatus {
    if (json == null) {
        return json;
    }
    return {
        
        'sessionId': json['session_id'],
        'state': json['state'],
        'seq': json['seq'],
        'workerId': json['worker_id'] == null ? undefined : json['worker_id'],
        'reason': json['reason'] == null ? undefined : json['reason'],
        'detail': json['detail'] == null ? undefined : json['detail'],
        'updatedAt': (new Date(json['updated_at'])),
    };
}

export function SessionStatusToJSON(json: any): SessionStatus {
    return SessionStatusToJSONTyped(json, false);
}

export function SessionStatusToJSONTyped(value?: SessionStatus | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'session_id': value['sessionId'],
        'state': value['state'],
        'seq': value['seq'],
        'worker_id': value['workerId'],
        'reason': value['reason'],
        'detail': value['detail'],
        'updated_at': value['updatedAt'].toISOString(),
    };
}